MVC. Architectural code refactoring. Complete rewrite. In progress.
